SCUMM: Fix hanging Maniac Mansion cutscene with the French version (Trac#13473)
The lab cutscene where Purple Tentacle is bullying Sandy would hang
when Dr Fred is done talking, in the official French version.

This appears to be related to this VAR_CHARCOUNT check in script no. 155:

[0085] (14) print(9,"{{Dr Fred's reaction line, shorter in French}}");
[00DA] (80) breakHere();
[00DB] (44) unless (VAR_CHARCOUNT > 90) goto 00DA;

Usually, French sentences are a bit longer than English sentences, but
in this case, it's shorter, but yet the script didn't adjust the 90
value, so it would indefinitely loop, causing the scene to hang (until
the player presses the Esc key) and Sandy's reaction to be lost.

So we just pad this string with extra spaces if its length looks too
short for the default VAR_CHARCOUNT check.  This should also remain
harmless if the string *is* long enough, or if a translation did fix
the VAR_CHARCOUNT check.

Still, we only restrict this workaround to the French version for now,
since we're not aware of a similar problem with other translations.