MATH: Just look for a GCC compatible compiler for __builtin_clz()
There's no need to check for GCC 3.4+ anymore, since we always require
a C++11 compiler.  Just look for __GNUC__, since this is an old
builtin and Clang always has it too.

Also #undef the LT() temporary macro while there, so that it can't
pollute something else by accident, since it's a short macro name in
a header file.