BASE: Fix Remaining Macro Definition Warnings
These are emitted by GCC when -Wundef is passed.

This could have caused odd behaviour if the WIN32 symbol was defined
as 0 i.e. #ifdef WIN32 code is included for compilation, but #if WIN32
guarded code is excluded, hence #ifdef is used in all cases as per the
existing compile time guards.