TWINE: fixed regression introduced in 3d5d5317a8
the modification of the processActor values from checkActorCollisionWithBricks and checkHeroCollisionWithBricks were
lost. The bug is that you could move crates into walls in the 'sokoban'-scene (scene 35).

See https://bugs.scummvm.org/ticket/13177

Once we remove the processActorSave hack, this would re-introduced the tank bug in scene 63 - that's why the fix
is non functional yet.