BLADERUNNER: Gracefully handle a case for Grigorian's arrest
Addresses bug #13727 which only happens in vanilla mode, and only when CDFRAMESx files are used instead of a single HDFRAMES

This bug can happen since by (my) mistake the kClueGrigoriansResources clue could be acquired in vanilla mode in ScummVM 2.5.0, 2.5.1 and 2.6.0.
That clue is supposed to only be part of the restored content mode.