SCUMM: Fix a continuity error with George Washington's teeth (DOTT)
You can't make George Washington leave the main hall without starting a
fire, but this fire can't be started if George Washington's is not
wearing the chattering teeth.

So, when he comes back to that room, he should be wearing them, but
this is not the case during the associated cutscene (he does wear them
if you talk to him outside, or come back to this room again after this
cutscene).  So, force costume 69 (chattering teeth) instead of costume
53 (no chattering teeth) for George Washington in this script.