SCUMM: Work around typo in Syd's biography in DOS Maniac Mansion (v1)
In the English version, Syd's biography says that he's "tring" to start
his own new-wave band. Other versions I've seen (in person or on YouTube)
have either fixed the typo, or (in the case of the NES version)
completely rewritten the text. Fix this, and mark it as an enhancement.